fix(native-app): fix Android back button navigation bug and inbox screen refactor.
